pysec_api/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Bcrypt
PyJWT
python-dotenv
pytest
pytest-flask
